+++ core/camel-core/src/main/java/org/apache/camel/builder/ExpressionClauseSupport.java
@@ -20,6 +20,7 @@ import java.util.Map;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Expression;
+import org.apache.camel.ExpressionFactory;
 import org.apache.camel.model.language.ConstantExpression;
 import org.apache.camel.model.language.ExchangePropertyExpression;
 import org.apache.camel.model.language.GroovyExpression;
@@ -39,7 +40,6 @@ import org.apache.camel.model.language.XMLTokenizerExpression;
 import org.apache.camel.model.language.XPathExpression;
 import org.apache.camel.model.language.XQueryExpression;
 import org.apache.camel.support.builder.xml.Namespaces;
-import org.apache.camel.support.language.ExpressionModel;
 
 /**
  * A support class for building expression clauses.
@@ -51,7 +51,7 @@ public class ExpressionClauseSupport<T> {
 
     private T result;
     private Expression expressionValue;
-    private ExpressionModel expressionType;
+    private ExpressionFactory expressionType;
 
     public ExpressionClauseSupport(T result) {
         this.result = result;
@@ -65,16 +65,16 @@ public class ExpressionClauseSupport<T> {
      */
     public T expression(Expression expression) {
         setExpressionValue(expression);
-        if (expression instanceof ExpressionModel) {
-            setExpressionType((ExpressionModel) expression);
+        if (expression instanceof ExpressionFactory) {
+            setExpressionType((ExpressionFactory) expression);
         }
         return result;
     }
 
     /**
-     * Specify an {@link ExpressionModel} instance
+     * Specify an {@link ExpressionFactory} instance
      */
-    public T language(ExpressionModel expression) {
+    public T language(ExpressionFactory expression) {
         setExpressionType(expression);
         return result;
     }
@@ -1087,11 +1087,11 @@ public class ExpressionClauseSupport<T> {
         this.expressionValue = expressionValue;
     }
 
-    public ExpressionModel getExpressionType() {
+    public ExpressionFactory getExpressionType() {
         return expressionType;
     }
 
-    public void setExpressionType(ExpressionModel expressionType) {
+    public void setExpressionType(ExpressionFactory expressionType) {
         this.expressionType = expressionType;
     }
